+++ backend/src/graphql/models/GdtSumPerEmail.ts
@@ -1,0 +1,19 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+/*
+import { ObjectType, Field } from 'type-graphql'
+
+@ObjectType()
+export class GdtSumPerEmail {
+  constructor(email: string, summe: number) {
+    this.email = email
+    this.summe = summe
+  }
+
+  @Field(() => String)
+  email: string
+
+  @Field(() => Number)
+  summe: number
+}
+*/
