+++ backend/src/graphql/models/Transaction.ts
@@ -53,32 +53,3 @@ export class Transaction {
   @Field({ nullable: true })
   decay?: Decay
 }
-
-@ObjectType()
-export class TransactionList {
-  constructor() {
-    this.gdtSum = 0
-    this.count = 0
-    this.balance = 0
-    this.decay = 0
-    this.decayDate = ''
-  }
-
-  @Field(() => Number)
-  gdtSum: number
-
-  @Field(() => Number)
-  count: number
-
-  @Field(() => Number)
-  balance: number
-
-  @Field(() => Number)
-  decay: number
-
-  @Field(() => String)
-  decayDate: string
-
-  @Field(() => [Transaction])
-  transactions: Transaction[]
-}
